AllFine

+++ Other/diamond2.cpp
@@ -5,28 +5,28 @@
 using namespace std;
 
 struct V {
-  int l, d[256];
+  LL l, d[256];
 } kMul, kMask, x, z, _i, _j, s[6], _s[3];
 string str;
 
 void M(V &v) {
   v.l = str.length();
   fill(&v.d[0], &v.d[255] + 1, 0);
-  for (int i = 0; i < v.l; i++) {
+  for (LL i = 0; i < v.l; i++) {
     v.d[v.l - i - 1] = str[i] - '0';
   }
 }
 
 string _M(V v) {
   string s = "";
-  for (int i = 0; i < v.l; i++) {
+  for (LL i = 0; i < v.l; i++) {
     s += (v.d[v.l - i - 1] + '0');
   }
   return s;
 }
 
 V C(V a) {
-  for (int i = 0; i < a.l; i++) {
+  for (LL i = 0; i < a.l; i++) {
     a.d[i + 1] += a.d[i] / 10, a.d[i] %= 10, a.l += a.d[a.l] > 0;
   }
   while (a.d[a.l - 1] == 0) {
@@ -40,7 +40,7 @@ V Add(V a, V b) {
     swap(a, b);
   }
   V c = {a.l, {}};
-  for (int i = 0; i < a.l; i++) {
+  for (LL i = 0; i < a.l; i++) {
     c.d[i] = a.d[i] + b.d[i];
   }
   return C(c);
@@ -51,8 +51,8 @@ V Mul(V a, V b) {
     swap(a, b);
   }
   V c = {a.l + b.l, {}};
-  for (int i = 0; i < b.l; i++) {
-    for (int j = 0; j < a.l; j++) {
+  for (LL i = 0; i < b.l; i++) {
+    for (LL j = 0; j < a.l; j++) {
       c.d[i + j] += a.d[j] * b.d[i];
     }
   }
@@ -64,7 +64,7 @@ V Xor(V a, V b) {
     swap(a, b);
   }
   V c = {a.l, {}};
-  for (int i = 0; i < a.l; i++) {
+  for (LL i = 0; i < a.l; i++) {
     c.d[i] = a.d[i] ^ b.d[i];
   }
   return C(c);
@@ -75,15 +75,15 @@ V And(V a, V b) {
     swap(a, b);
   }
   V c = {a.l, {}};
-  for (int i = 0; i < a.l; i++) {
+  for (LL i = 0; i < a.l; i++) {
     c.d[i] = a.d[i] & b.d[i];
   }
   return C(c);
 }
 
-V Sh(V a, int d) {  //positive:LSH negative:RSH
+V Sh(V a, LL d) {  //positive:LSH negative:RSH
   V c = {a.l, {}};
-  for (int i = 0; i < a.l; i++) {
+  for (LL i = 0; i < a.l; i++) {
     c.d[i] = d >= 0 ? a.d[i] << d : a.d[i] >> (-d);
   }
   return C(c);
@@ -98,16 +98,16 @@ int main() {
   cin >> str, M(x);
   cin >> str, M(z);
   s[1] = And(Xor(s[0], kMul), kMask);
-  for (int i = 2; i <= 5; i++) {
+  for (LL i = 2; i <= 5; i++) {
     s[i] = And(Add(Mul(s[i - 1], kMul), {2, {1, 1}}), kMask);
   }
-  for (int i = 2; i <= 5; i++) {
+  for (LL i = 2; i <= 5; i++) {
     s[i] = i & 1 ? Sh(Sh(s[i], 16), -32) : Sh(Sh(s[i], -16), 32);
   }
   _i = Add(s[2], s[3]), _j = Add(s[4], s[5]);
   _i.d[_i.l - 1] += !(_i.d[_i.l - 1] & 1), _j.d[_j.l - 1] += !(_j.d[_j.l - 1] & 1);
   _s[0] = And(Xor(Add(Xor(Add(Mul(Mul({2, {6, 1}}, x), _i), Mul(Mul({2, {6, 1}}, z), _j)), s[0]), {5, {9, 0, 0, 0, 6}}), kMul), kMask);
-  for (int i = 1; i <= 2; i++) {
+  for (LL i = 1; i <= 2; i++) {
     _s[i] = And(Add(Mul(_s[i - 1], kMul), {2, {1, 1}}), kMask);
   }
   cout << "The diamond is at x:" << _M(Add(Sh(_s[1], -44), Mul({2, {6, 1}}, x))) << " z:" << _M(Add(Sh(_s[2], -44), Mul({2, {6, 1}}, z))) << endl;
